+++ src/headless/converse-muc.js
@@ -230,14 +230,14 @@ converse.plugins.add('converse-muc', {
             }
         }
 
-        function openChatRoom (jid, settings) {
+        async function openChatRoom (jid, settings) {
             /* Opens a groupchat, making sure that certain attributes
              * are correct, for example that the "type" is set to
              * "chatroom".
              */
             settings.type = _converse.CHATROOMS_TYPE;
             settings.id = jid;
-            const chatbox = _converse.chatboxes.getChatBox(jid, settings, true);
+            const chatbox = await _converse.chatboxes.getChatBox(jid, settings, true);
             chatbox.maybeShow(true);
             return chatbox;
         }
@@ -2040,7 +2040,7 @@ converse.plugins.add('converse-muc', {
          * @method _converse.ChatRoom#onDirectMUCInvitation
          * @param { XMLElement } message - The message stanza containing the invitation.
          */
-        _converse.onDirectMUCInvitation = function (message) {
+        _converse.onDirectMUCInvitation = async function (message) {
             const x_el = sizzle('x[xmlns="jabber:x:conference"]', message).pop(),
                 from = Strophe.getBareJidFromJid(message.getAttribute('from')),
                 room_jid = x_el.getAttribute('jid'),
@@ -2066,7 +2066,7 @@ converse.plugins.add('converse-muc', {
                 }
             }
             if (result === true) {
-                const chatroom = openChatRoom(room_jid, {'password': x_el.getAttribute('password') });
+                const chatroom = await openChatRoom(room_jid, {'password': x_el.getAttribute('password') });
                 if (chatroom.get('connection_status') === converse.ROOMSTATUS.DISCONNECTED) {
                     _converse.chatboxes.get(room_jid).join();
                 }
@@ -2196,6 +2196,7 @@ converse.plugins.add('converse-muc', {
                  * @param {(string[]|string)} jid|jids The JID or array of
                  *     JIDs of the chatroom(s) to create
                  * @param {object} [attrs] attrs The room attributes
+                 * @returns {Promise} Promise which resolves with the Backbone.Model representing the chat.
                  */
                 create (jids, attrs={}) {
                     attrs = _.isString(attrs) ? {'nick': attrs} : (attrs || {});
@@ -2239,6 +2240,7 @@ converse.plugins.add('converse-muc', {
                  *   another chat already in the foreground.
                  *   Set `force` to `true` if you want to force the room to be
                  *   maximized or shown.
+                 * @returns {Promise} Promise which resolves with the Backbone.Model representing the chat.
                  *
                  * @example
                  * this._converse.api.rooms.open('group@muc.example.com')
@@ -2275,13 +2277,14 @@ converse.plugins.add('converse-muc', {
                         _converse.log(err_msg, Strophe.LogLevel.ERROR);
                         throw(new TypeError(err_msg));
                     } else if (_.isString(jids)) {
-                        const room = _converse.api.rooms.create(jids, attrs);
-                        if (room) {
-                            room.maybeShow(force);
-                        }
+                        const room = await _converse.api.rooms.create(jids, attrs);
+                        room && room.maybeShow(force);
                         return room;
                     } else {
-                        return jids.map(jid => _converse.api.rooms.create(jid, attrs).maybeShow(force));
+                        return jids.map(async jid => {
+                            const room = await _converse.api.rooms.create(jid, attrs);
+                            room.maybeShow(force);
+                        });
                     }
                 },
 
@@ -2298,6 +2301,7 @@ converse.plugins.add('converse-muc', {
                  *     the user's JID will be used.
                  * @param {boolean} create A boolean indicating whether the room should be created
                  *     if not found (default: `false`)
+                 * @returns {Promise} Promise which resolves with the Backbone.Model representing the chat.
                  * @example
                  * _converse.api.waitUntil('roomsAutoJoined').then(() => {
                  *     const create_if_not_found = true;